requirements.txt
alabaster==0.7.13
apparmor==3.1.2
appdirs==1.4.4
async-generator==1.10
attrs==22.2.0
autocommand==2.2.2
Babel==2.11.0
beautifulsoup4==4.11.1
Brlapi==0.8.4
btrfsutil==6.1.2
build==0.10.0
cairocffi==1.4.0
CairoSVG==2.5.2
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.0.7
cryptography==38.0.4
cssselect==1.2.0
cssselect2==0.7.0
cycler==0.11.0
Cython==0.29.33
defusedxml==0.7.1
distro==1.8.0
docutils==0.19
evdev==1.6.1
exceptiongroup==1.0.4
fastjsonschema==2.16.2
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.29
guiscrcpy==2023.1.1
h11==0.14.0
humanfriendly==10.0
idna==3.4
imagesize==1.4.1
inflect==6.0.2
installer==0.6.0
jaraco.context==4.2.0
jaraco.functools==3.5.2
jaraco.text==3.11.0
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
lark==1.1.5
lensfun==0.3.3
LibAppArmor==3.1.2
libfdt==1.6.1
louis==3.24.0
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
mercurial==6.3.2
meson==1.0.0
more-itertools==9.0.0
mysql-connector==2.2.9
numpy==1.24.0
ordered-set==4.1.0
outcome==1.2.0
packaging==21.3
pandas==1.5.3
pep517==0.13.0
Pillow==9.4.0
Pivy==0.6.8
ply==3.11
poetry-core==1.4.0
psutil==5.9.4
pwquality==1.4.5
pybind11==2.10.3
pycparser==2.21
pydantic==1.10.4
Pygments==2.14.0
PyGObject==3.42.2
PyNaCl==1.4.0
pynput==1.7.6
pyparsing==3.0.9
pyproject_hooks==1.0.0
PyQt5==5.15.7
PyQt5-sip==12.11.0
pyrsistent==0.19.3
PySide2==5.15.8
PySocks==1.7.1
python-dateutil==2.8.2
python-xlib==0.31
pytz==2022.7
PyYAML==6.0
QtPy==2.3.0
Reflector==2021.11.20.2.41.3
requests==2.28.1
requests-futures==1.0.0
selenium==4.7.2
shiboken2==5.15.8
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
Sphinx==5.3.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stem==1.8.0
TBB==0.2
team==1.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
torrequest==0.1.0
trio==0.22.0
trio-websocket==0.9.2
trove-classifiers==2023.1.20
typing_extensions==4.4.0
urllib3==1.26.12
validate-pyproject==0.11
webencodings==0.5.1
wsproto==1.2.0
xcffib==1.1.0
youtube-dl==2021.12.17
